requirements.txt
flask
openai
pandas
pinecone
python-dotenv
tiktoken